+++ scripts/test.sh
@@ -3,10 +3,10 @@
 # setparams
 # 1 day   = 86400
 # 30 days = 2592000
-cleos push action eosio.wps setparams '[{"vote_margin": 2, "deposit_required": "100.0000 EOS", "voting_interval": 120, "max_monthly_budget": "1000.0000 EOS", "min_time_voting_end": 10}]' -p eosio.wps
+cleos push action eosio.wps setparams '[{"vote_margin": 2, "deposit_required": "100.0000 EOS", "voting_interval": 2592000, "max_monthly_budget": "1000.0000 EOS", "min_time_voting_end": 86400}]' -p eosio.wps
 
 # init
-cleos -v push action eosio.wps init '["1970-01-01T00:00:00"]' -p eosio.wps
+cleos -v push action eosio.wps init '["2019-12-29T00:00:00"]' -p eosio.wps
 
 # set proposer
 cleos -v push action eosio.wps setproposer '["myaccount", [{"key":"region", "value":"global"}]]' -p myaccount
@@ -16,6 +16,7 @@ cleos -v push action eosio.wps submitdraft '["myaccount", "mywps", "My WPS", "50
 cleos -v push action eosio.wps submitdraft '["toaccount", "towps", "To WPS", "200.0000 EOS", 6, [{"key":"category", "value":"other"}]]' -p toaccount
 cleos -v push action eosio.wps submitdraft '["toaccount", "short", "Short WPS", "800.0000 EOS", 1, [{"key":"category", "value":"short"}]]' -p toaccount
 cleos -v push action eosio.wps submitdraft '["toaccount", "novote", "No Vote WPS", "500.0000 EOS", 1, [{"key":"category", "value":"novote"}]]' -p toaccount
+cleos -v push action eosio.wps submitdraft '["toaccount", "optional", "Optional Vote WPS", "500.0000 EOS", 1, [{"key":"category", "value":"optional"}]]' -p toaccount
 
 # # cancel draft
 # cleos -v push action eosio.wps canceldraft '["myaccount", "mywps"]' -p myaccount
@@ -31,13 +32,13 @@ cleos -v transfer toaccount eosio.wps "400.0000 EOS" ""
 # cleos -v push action eosio.wps refund '["myaccount"]' -p myaccount
 
 # activate
-cleos -v push action eosio.wps activate '["myaccount", "mywps", "1970-01-01T00:00:00"]' -p myaccount
-cleos -v push action eosio.wps activate '["toaccount", "towps", "1970-01-01T00:00:00"]' -p toaccount
-cleos -v push action eosio.wps activate '["toaccount", "short", "1970-01-01T00:00:00"]' -p toaccount
-cleos -v push action eosio.wps activate '["toaccount", "novote", "1970-01-01T00:00:00"]' -p toaccount
+cleos -v push action eosio.wps activate '["myaccount", "mywps", null]' -p myaccount
+cleos -v push action eosio.wps activate '["toaccount", "towps", null]' -p toaccount
+cleos -v push action eosio.wps activate '["toaccount", "short", null]' -p toaccount
+cleos -v push action eosio.wps activate '["toaccount", "novote", null]' -p toaccount
 
 # fund wps
-cleos -v transfer eosio.ramfee eosio.wps "2000.0000 EOS" ""
+cleos -v transfer eosio.ramfee eosio.wps "50000.0000 EOS" ""
 
 # # cancel
 # cleos -v push action eosio.wps canceldraft '["myaccount", "mywps"]' -p myaccount